+++ demos/Tuio2Demo.cpp
@@ -263,7 +263,7 @@ int main(int argc, char* argv[])
 		return 0;
 	}
 
-	int port = 3333;
+	unsigned short port = 3333;
 	if( argc >= 2 ) port = atoi( argv[1] );
 
 	Tuio2Demo *app = new Tuio2Demo(port);
